requirements.txt
matplotlib>=2.2.2
numpy>=1.14.5
Pillow>=5.1.0
scipy>=1.1.0
seaborn>=0.8.1
torch>=0.4.1
torchvision>=0.2.1
